update skema database

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('username')->unique();
             $table->string('password');
             $table->enum('role', ['admin', 'pembimbing_internal']);
-            $table->unsignedBigInteger('id_ref')->nullable();
+            $table->unsignedInteger('id_ref')->nullable();
             $table->foreign('id_ref')->references('id_pembimbing_i')->on('pembimbing_internal')->nullOnDelete();
             $table->timestamps();
         });
